+++ src/generate_system/scalar_eqs.jl
@@ -25,7 +25,6 @@ function scalar_eqs!(
     R_v_to_w, pos
 )
     (; wings) = s.sys_struct
-    wind_factor = param_computed!(params.reg, :wind_factor, WindFactorReader())
     @variables begin
         # Body frame axes and apparent wind (column-major: [1:3, wing_idx])
         e_x(t)[1:3, eachindex(wings)]
@@ -43,7 +42,8 @@ function scalar_eqs!(
             e_y[:, wing.idx] ~ R_b_to_w[:, 2, wing.idx]
             e_z[:, wing.idx] ~ R_b_to_w[:, 3, wing.idx]
             wind_vel_wing[:, wing.idx] ~
-                wind_factor(wing_pos[3, wing.idx]) * wind_vec_gnd
+                wing_wind_source(params, wing.idx, wind_vec_gnd)(
+                    wing_pos[3, wing.idx])
             wind_disturb[:, wing.idx] ~ params.wings[wing.idx].wind_disturb
             va_wing[:, wing.idx] ~
                 wind_vel_wing[:, wing.idx] - wing_vel[:, wing.idx] +